Turbulox

Documentation for Turbulox.

Turbulox.Turbulox

Turbulox.Grid

Turbulox.apply!

Turbulox.convectiondiffusion!

Turbulox.convterm

Turbulox.default_right_hand_side!

Turbulox.dim

Turbulox.divergence!

Turbulox.e

Turbulox.getval

Turbulox.order

Turbulox.poissonsolver

Turbulox.pressuregradient!

Turbulox.problem_setup

Turbulox.project!

Turbulox.propose_timestep

Turbulox.scalarfield

Turbulox.tensorfield

Turbulox.timestep!

Turbulox.vectorfield

Turbulox.Turbulox — Module

Simulate turbulence in a box.

Turbulox.Grid — Type

Staggered grid of order o and dimension d.

Turbulox.apply! — Method

Apply kernel! on setup.grid, args... over the entire domain. The args are typically input and output fields. The kernel should be of the form

using KernelAbstractions
@kernel function kernel!(grid, args...)
    # content
end

Turbulox.convectiondiffusion! — Function

Compute convection-diffusion force from velocity u. Add the force field to f.

Turbulox.convterm — Function

Approximate the convective force $\partial_j (u_i u_j)$.

Turbulox.default_right_hand_side! — Method

Deault right-hand side function (without projection).

Turbulox.dim — Method

Get physical dimension.

Turbulox.divergence! — Function

Compute divergence of vector field u. Put the result in div.

Turbulox.e — Method

Get unit index in dimension α.

Turbulox.getval — Method

Get value from Val.

Turbulox.order — Method

Get order of grid.

Turbulox.poissonsolver — Method

Create spectral Poisson solver from setup.

Turbulox.pressuregradient! — Function

Subtract pressure gradient.

Turbulox.problem_setup — Method

Problem setup.

Kwargs

grid::Grid: Grid setup.

visc::Real: Viscosity. This value is also used to infer the floating point type, so make sure it is a Float32 or Float64.

backend = CPU(): KernelAbstractions.jl backend. For Nvidia GPUs, do using CUDA and set to CUDABackend().

workgroupsize = 64: Kernel work group size.

Turbulox.project! — Method

Project velocity field onto divergence-free space.

Turbulox.propose_timestep — Method

Get proposed maximum time step for convection and diffusion terms.

Turbulox.scalarfield — Method

Allocate empty scalar field.

Turbulox.tensorfield — Method

Allocate empty tensor field.

Turbulox.timestep! — Method

Perform time step using Wray's third-order scheme.

Turbulox.vectorfield — Method

Allocate empty vector field.