Turbulox

Documentation for Turbulox.

Turbulox.apply!

Turbulox.convectiondiffusion!

Turbulox.default_right_hand_side!

Turbulox.divergence!

Turbulox.getval

Turbulox.periodicindex

Turbulox.poissonsolver

Turbulox.pressuregradient!

Turbulox.problem_setup

Turbulox.project!

Turbulox.scalarfield

Turbulox.tensorfield

Turbulox.timestep!

Turbulox.unitindices

Turbulox.vectorfield

Turbulox.apply! — Method

Apply kernel! on setup, args... over the entire domain. The args are typically input and output fields. The kernel should be of the form

using KernelAbstractions
@kernel function kernel!(setup, args...)
    # content
end

Turbulox.convectiondiffusion! — Function

Compute convection-diffusion force from velocity u. Add the force field to f.

Turbulox.default_right_hand_side! — Method

Deault right-hand side function (without projection).

Turbulox.divergence! — Function

Compute divergence of vector field u. Put the result in div.

Turbulox.getval — Method

Get value from Val.

Turbulox.periodicindex — Function

Extend index periodically so that it stays within the domain.

Turbulox.poissonsolver — Method

Create spectral Poisson solver from setup.

Turbulox.pressuregradient! — Function

Subtract pressure gradient.

Turbulox.problem_setup — Method

Problem setup.

Kwargs

D = Val(2): Physical dimension. This is wrapped in a Val to make it known at compile time.

n::Int: Number of grid points in each dimension.

visc::Real: Viscosity. This value is also used to infer the floating point type, so make sure it is a Float32 or Float64.

backend = CPU(): KernelAbstractions.jl backend. For Nvidia GPUs, do using CUDA and set to CUDABackend().

Turbulox.project! — Method

Project velocity field onto divergence-free space.

Turbulox.scalarfield — Method

Allocate empty scalar field.

Turbulox.tensorfield — Method

Allocate empty tensor field.

Turbulox.timestep! — Method

Perform time step using Wray's third-order scheme.

Turbulox.unitindices — Function

Get unit indices for given dimension.

Turbulox.vectorfield — Method

Allocate empty vector field.