Turbulox

Documentation for Turbulox.

Turbulox.Turbulox

Turbulox.Grid

Turbulox.apply!

Turbulox.clark_model

Turbulox.collocated_tensorfield

Turbulox.convectiondiffusion!

Turbulox.convterm

Turbulox.default_right_hand_side!

Turbulox.dim

Turbulox.divergence!

Turbulox.e

Turbulox.eddyviscosity_closure!

Turbulox.eddyviscosity_model

Turbulox.get_scale_numbers

Turbulox.getval

Turbulox.nicoud_viscosity!

Turbulox.order

Turbulox.poissonsolver

Turbulox.pol_tensor_collocated

Turbulox.pressuregradient!

Turbulox.problem_setup

Turbulox.project!

Turbulox.propose_timestep

Turbulox.s3pqr_viscosity!

Turbulox.scalarfield

Turbulox.smagorinsky_viscosity!

Turbulox.staggered_tensorfield

Turbulox.tensordivergence!

Turbulox.tensordivergence_collocated!

Turbulox.timestep!

Turbulox.vectorfield

Turbulox.verstappen_viscosity!

Turbulox.vreman_viscosity!

Turbulox.Turbulox — Module

Simulate turbulence in a box.

Turbulox.Grid — Type

Staggered grid of order o and dimension d.

Turbulox.apply! — Method

Apply kernel! on setup.grid, args... over the entire domain. The args are typically input and output fields. The kernel should be of the form

using KernelAbstractions
@kernel function kernel!(grid, args...)
    # content
end

Turbulox.clark_model — Method

Clark's gradient model [1].

Turbulox.collocated_tensorfield — Method

Allocate empty tensor field (collocated).

Turbulox.convectiondiffusion! — Function

Compute convection-diffusion force from velocity u. Add the force field to f.

Turbulox.convterm — Function

Approximate the convective force $\partial_j (u_i u_j)$.

Turbulox.default_right_hand_side! — Method

Default right-hand side function (without projection).

Turbulox.dim — Method

Get physical dimension.

Turbulox.divergence! — Function

Compute divergence of vector field u. Put the result in div.

Turbulox.e — Method

Get unit index in dimension i.

Turbulox.eddyviscosity_closure! — Method

Divergence of 2 * visc * S. Interpolate visc to staggered points first. Add result to existing force field f.

Turbulox.eddyviscosity_model — Method

Eddy viscosity closure model.

Turbulox.get_scale_numbers — Method

Get the following dimensional scale numbers [2]:

Velocity $u_\text{avg} = \langle u_i u_i \rangle^{1/2}$

Dissipation rate $\epsilon = 2 \nu \langle S_{ij} S_{ij} \rangle$

Kolmolgorov length scale $\eta = (\frac{\nu^3}{\epsilon})^{1/4}$

Taylor length scale $\lambda = (\frac{5 \nu}{\epsilon})^{1/2} u_\text{avg}$

Taylor-scale Reynolds number $Re_\lambda = \frac{\lambda u_\text{avg}}{\sqrt{3} \nu}$

Integral length scale $L = \frac{3 \pi}{2 u_\text{avg}^2} \int_0^\infty \frac{E(k)}{k} \, \mathrm{d} k$

Large-eddy turnover time $\tau = \frac{L}{u_\text{avg}}$

Turbulox.getval — Method

Get value from Val.

Turbulox.nicoud_viscosity! — Method

Compute Nicoud's eddy viscosity ($\sigma$-model) [3, 4] Proposed value for C is 1.35.

Turbulox.order — Method

Get order of grid.

Turbulox.poissonsolver — Method

Create spectral Poisson solver from setup.

Turbulox.pol_tensor_collocated — Method

Interpolate staggered tensor to volume centers.

Turbulox.pressuregradient! — Function

Subtract pressure gradient.

Turbulox.problem_setup — Method

Problem setup.

Kwargs

grid::Grid: Grid setup.

visc::Real: Viscosity. This value is also used to infer the floating point type, so make sure it is a Float32 or Float64.

backend = CPU(): KernelAbstractions.jl backend. For Nvidia GPUs, do using CUDA and set to CUDABackend().

workgroupsize = 64: Kernel work group size.

Turbulox.project! — Method

Project velocity field onto divergence-free space.

Turbulox.propose_timestep — Method

Get proposed maximum time step for convection and diffusion terms.

Turbulox.s3pqr_viscosity! — Method

Compute S3PQR eddy viscosity [4]

Proposed values for $p$ (set valp = Val(p)):

-5/2

-1

0

Turbulox.scalarfield — Method

Allocate empty scalar field.

Turbulox.smagorinsky_viscosity! — Method

Compute Smagorinsky's original eddy viscosity [5]. Proposed value for C is 0.17.

Turbulox.staggered_tensorfield — Method

Allocate empty tensor field (staggered).

Turbulox.tensordivergence! — Method

Divergence of staggered tensor field. Add result to existing force field f.

Turbulox.tensordivergence_collocated! — Method

Divergence of collocated tensor field. First interpolate to staggered points. Add result to existing force field f.

Turbulox.timestep! — Method

Perform time step using Wray's third-order scheme.

Turbulox.vectorfield — Method

Allocate empty vector field.

Turbulox.verstappen_viscosity! — Method

Compute Verstappen's eddy viscosity [4, 6]. Proposed values for C are 0.345 [6] or 0.527 [4]."

Turbulox.vreman_viscosity! — Method

Compute Vreman's eddy viscosity [4, 7]. Proposed value for C is 0.28.