Turbulox

Documentation for Turbulox.

Turbulox.Turbulox

Turbulox.Coll

Turbulox.Grid

Turbulox.Stag

Turbulox.apply!

Turbulox.applyfilter!

Turbulox.applyfilter!

Turbulox.applyfilter!

Turbulox.clark_model

Turbulox.collocated_tensorfield

Turbulox.convectiondiffusion!

Turbulox.convterm

Turbulox.default_right_hand_side!

Turbulox.desymmetrize!

Turbulox.deviator

Turbulox.dim

Turbulox.divergence!

Turbulox.dx

Turbulox.e

Turbulox.eddyviscosity_closure!

Turbulox.eddyviscosity_model

Turbulox.expand_tensorbasis!

Turbulox.fill_invariants!

Turbulox.fill_tensorbasis!

Turbulox.gaussian

Turbulox.get_axis

Turbulox.get_scale_numbers

Turbulox.getval

Turbulox.mergestencil

Turbulox.nicoud_viscosity!

Turbulox.order

Turbulox.poissonsolver

Turbulox.pol_tensor_collocated

Turbulox.pol_tensor_stag

Turbulox.pressuregradient!

Turbulox.project!

Turbulox.propose_timestep

Turbulox.s3pqr_viscosity!

Turbulox.scalarfield

Turbulox.smagorinsky_viscosity!

Turbulox.staggered_tensorfield

Turbulox.stress

Turbulox.surfacefilter_tensor!

Turbulox.symmetric_staggered_tensorfield

Turbulox.tensorbasis

Turbulox.tensordivergence!

Turbulox.tensordivergence_collocated!

Turbulox.tensordivergence_collocated_2!

Turbulox.tensorproduct_coll!

Turbulox.tensorproduct_stag!

Turbulox.timestep!

Turbulox.vectorfield

Turbulox.verstappen_viscosity!

Turbulox.volumefilter_tensor!

Turbulox.vreman_viscosity!

Turbulox.wale_viscosity!

Turbulox.Turbulox — Module

Simulate turbulence in a box.

Turbulox.Coll — Type

Collocated grid position.

Turbulox.Grid — Type

Staggered grid of order o and dimension d.

Turbulox.Stag — Type

Staggered grid position.

Turbulox.apply! — Method

Apply kernel! on grid, args... over the entire domain. The args are typically input and output fields. The kernel should be of the form

using KernelAbstractions
@kernel function kernel!(grid, args...)
    # content
end

Turbulox.applyfilter! — Method

Filter scalar field.

Turbulox.applyfilter! — Method

Filter staggered tensor field.

Turbulox.applyfilter! — Method

Filter staggered vector field.

Turbulox.clark_model — Method

Clark's gradient model [1].

Turbulox.collocated_tensorfield — Method

Allocate empty tensor field (collocated).

Turbulox.convectiondiffusion! — Function

Compute convection-diffusion force from velocity u. Add the force field to f.

Turbulox.convterm — Function

Approximate the convective force $\partial_j (u_i u_j)$.

Turbulox.default_right_hand_side! — Method

Default right-hand side function (without projection).

Turbulox.desymmetrize! — Method

Copy symmtric tensor to normal tensor

Turbulox.deviator — Method

Compute deviatoric part of a tensor.

Turbulox.dim — Method

Get physical dimension.

Turbulox.divergence! — Function

Compute divergence of vector field u. Put the result in div.

Turbulox.dx — Method

Get grid spacing.

Turbulox.e — Method

Get unit index in dimension i.

Turbulox.eddyviscosity_closure! — Method

Divergence of -2 * visc * S. Interpolate visc to staggered points first. Subtract result from existing force field f.

Turbulox.eddyviscosity_model — Method

Eddy viscosity closure model.

Turbulox.expand_tensorbasis! — Method

Expand tensor basis B and fill τ[x] = c[i, x] * B[i, x] (sum over i).

Turbulox.fill_invariants! — Method

Fill V with invariants.

Turbulox.fill_tensorbasis! — Method

Fill B with basis tensors.

Turbulox.gaussian — Method

Gaussian filter kernel.

Turbulox.get_axis — Method

Get grid points along axis.

Turbulox.get_scale_numbers — Method

Get the following dimensional scale numbers [2]:

Velocity $u_\text{avg} = \langle u_i u_i \rangle^{1/2}$

Dissipation rate $\epsilon = 2 \nu \langle S_{ij} S_{ij} \rangle$

Kolmolgorov length scale $\eta = (\frac{\nu^3}{\epsilon})^{1/4}$

Taylor length scale $\lambda = (\frac{5 \nu}{\epsilon})^{1/2} u_\text{avg}$

Taylor-scale Reynolds number $Re_\lambda = \frac{\lambda u_\text{avg}}{\sqrt{3} \nu}$

Integral length scale $L = \frac{3 \pi}{2 u_\text{avg}^2} \int_0^\infty \frac{E(k)}{k} \, \mathrm{d} k$

Large-eddy turnover time $\tau = \frac{L}{u_\text{avg}}$

Turbulox.getval — Method

Get value from Val.

Turbulox.mergestencil — Method

Merge stencil periodically if the stencil is longer than the grid size n.

Turbulox.nicoud_viscosity! — Method

Compute Nicoud's eddy viscosity ($\sigma$-model) [3, 4] Proposed value for C is 1.35.

Turbulox.order — Method

Get order of grid.

Turbulox.poissonsolver — Method

Create spectral Poisson solver from grid.

Turbulox.pol_tensor_collocated — Method

Interpolate staggered tensor to volume centers.

Turbulox.pol_tensor_stag — Method

Interpolate collocated tensor to staggered tensor.

Turbulox.pressuregradient! — Function

Subtract pressure gradient.

Turbulox.project! — Method

Project velocity field onto divergence-free space.

Turbulox.propose_timestep — Method

Get proposed maximum time step for convection and diffusion terms.

Turbulox.s3pqr_viscosity! — Method

Compute S3PQR eddy viscosity [4]

Proposed values for $p$ (set valp = Val(p)):

-5/2

-1

0

Turbulox.scalarfield — Method

Allocate empty scalar field.

Turbulox.smagorinsky_viscosity! — Method

Compute Smagorinsky's original eddy viscosity [5]. Proposed value for C is 0.17.

Turbulox.staggered_tensorfield — Method

Allocate empty tensor field (staggered).

Turbulox.stress — Method

Get convection-diffusion stress tensor component i,j.

Turbulox.surfacefilter_tensor! — Method

Surface-average staggered tensor field.

Turbulox.symmetric_staggered_tensorfield — Method

Allocate empty tensor field (symmetric, staggered).

Turbulox.tensorbasis — Function

Compute Pope's tensor basis [2].

Turbulox.tensordivergence! — Method

Divergence of staggered tensor field $σ$. Subtract result from existing force field $f$. The operation is $f_i \leftarrow f_i - ∂_j σ_{i j}$.

Turbulox.tensordivergence_collocated! — Method

Divergence of collocated tensor field $\sigma$. First interpolate to staggered points. Subtract result from existing force field $f$. The operation is $f_i \leftarrow f_i - ∂_j σ_{i j}$.

Turbulox.tensordivergence_collocated_2! — Method

Divergence first, then interpolate

Turbulox.tensorproduct_coll! — Method

Compute $u v^T$ in the collocated points.

Turbulox.tensorproduct_stag! — Method

Compute $u v^T$ in the staggered points.

Turbulox.timestep! — Method

Perform time step using Wray's third-order scheme.

Turbulox.vectorfield — Method

Allocate empty vector field.

Turbulox.verstappen_viscosity! — Method

Compute Verstappen's eddy viscosity [4, 6]. Proposed values for C are 0.345 [6] or 0.527 [4]."

Turbulox.volumefilter_tensor! — Method

Volume-average staggered tensor field.

Turbulox.vreman_viscosity! — Method

Compute Vreman's eddy viscosity [4, 7]. Proposed value for C is 0.28.

Turbulox.wale_viscosity! — Method

Compute WALE [4, 8]. Proposed value for C is 0.569.